Add utility for simulating dataset with mask/psf.
